Remove binary files
