--- BCR_report_builder/src/application/Main.java
@@ -1,133 +1,0 @@
-package application;
-	
-import javafx.application.Application;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.stage.Stage;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.FlowPane;
-import javafx.scene.layout.GridPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontWeight;
-import javafx.scene.text.Text;
-
-
-public class Main extends Application {
-	@Override
-	public void start(Stage primaryStage) {
-		try {
-			primaryStage.setTitle("JavaFX");
-			
-			GridPane grid = new GridPane();
-			//grid.setPadding(new Insets(130, 125, 150, 125));
-			grid.setPadding(new Insets(20, 20, 20, 20));
-			grid.setAlignment(Pos.CENTER);
-			grid.setHgap(10);
-			grid.setVgap(10);
-			Text scenetitle = new Text("Welcome");
-			scenetitle.setFont(Font.font("Tahoma", FontWeight.NORMAL, 25));
-			
-			Button createReport = new Button("Create Report");
-			Button editReport = new Button("Edit Report");
-			Button deleteReport = new Button("Delete Report");
-			
-			grid.add(scenetitle, 0, 0);
-			grid.add(createReport, 1, 2, 2, 2);
-			grid.add(editReport, 3, 2, 2, 2);
-			grid.add(deleteReport, 5, 2, 2, 2);
-			
-			Scene scene = new Scene(grid, 600, 600);
-			scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
-			primaryStage.setScene(scene);
-			primaryStage.show();
-		} catch(Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	public static void main(String[] args) {
-		launch(args);
-	}
-}
-
-
-/*
- * 
- import javafx.application.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
-import javafx.scene.layout.GridPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.paint.Color;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontWeight;
-import javafx.scene.text.Text;
-import javafx.stage.Stage;
-
-public class JavafxTutorial extends Application {
-    public static void main(String[] args) {
-        launch(args);
-    }
-    
-    @Override
-    public void start(Stage primaryStage) {
-        primaryStage.setTitle("JavaFX Welcome");
-        
-        GridPane grid = new GridPane();
-        grid.setAlignment(Pos.CENTER);
-        grid.setHgap(10);
-        grid.setVgap(10);
-        grid.setPadding(new Insets(25, 25, 25, 25));
-        
-        Text scenetitle = new Text("Welcome");
-        scenetitle.setFont(Font.font("Tahoma", FontWeight.NORMAL, 20));
-        grid.add(scenetitle, 0, 0, 2, 1);
-
-        Label userName = new Label("User Name:");
-        grid.add(userName, 0, 1);
-
-        TextField userTextField = new TextField();
-        grid.add(userTextField, 1, 1);
-
-        Label pw = new Label("Password:");
-        grid.add(pw, 0, 2);
-
-        PasswordField pwBox = new PasswordField();
-        grid.add(pwBox, 1, 2);
-        
-        final Text actiontarget = new Text();
-        grid.add(actiontarget, 1, 5);
-        
-        Button signIn = new Button("Sign In");
-        signIn.setOnAction(new EventHandler<ActionEvent>(){
-        	@Override
-        	public void handle(ActionEvent e){
-        		actiontarget.setFill(Color.FIREBRICK);
-        		actiontarget.setText("Sign In Button Pressed");
-        	}
-        });
-        
-        HBox hbx = new HBox(10);
-        hbx.setAlignment(Pos.BOTTOM_RIGHT);
-        hbx.getChildren().add(signIn);
-        grid.add(hbx, 1, 3);
-        
-
-        
-        Scene scene = new Scene(grid, 300, 275);
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
-}
-
-*/
