--- BCR_report_builder/src/application/ReportBuilderException.java
@@ -1,16 +1,0 @@
-package application;
-
-public class ReportBuilderException extends Exception{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-		/*
-		 * TODO:
-		 * 		Create the necessary exception infrastructure
-		 * 		
-		 */
-
-}
