--- Gameoflife/src/cellPosition.java
@@ -1,9 +1,0 @@
-class cellPosition {
-	int row;
-	int column;
-
-	cellPosition(int row, int column){
-		this.row = row;
-		this.column = column;
-	}
-}
