+++ README.md
@@ -1,0 +1,23 @@
+# Game of Life
+
+This project uses Gradle and JavaFX 21. You need JDK 21 or later.
+
+## Building
+
+```bash
+./gradlew build
+```
+
+## Running
+
+```bash
+./gradlew run
+```
+
+The runnable JAR is created in `build/libs/game-of-life.jar`.
+
+## Testing
+
+```
+./gradlew test
+```
