+++ src/main/java/com/example/gameoflife/CellPosition.java
@@ -1,0 +1,10 @@
+package com.example.gameoflife;
+class CellPosition {
+	int row;
+	int column;
+
+	CellPosition(int row, int column){
+		this.row = row;
+		this.column = column;
+	}
+}
