+++ src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module gameoflife {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+    exports com.example.gameoflife;
+}
