+++ src/test/java/com/example/gameoflife/CellTest.java
@@ -1,3 +1,4 @@
+package com.example.gameoflife;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
